requirements.txt
langchain==0.1.16
langchain-community==0.0.32
faiss-cpu
elasticsearch
sentence-transformers
openai
